tests/wpt/web-platform-tests/css/css-grid/abspos/positioned-grid-items-should-not-take-up-space-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Positioned grid items should not take up space</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#abspos" title="9. Absolute Positioning">
<meta name="assert" content="This test checks that positioned items shouldn't take up space or otherwise participate in the layout of the grid.">
<link rel="stylesheet" href="/fonts/ahem.css">
<link href="/css/support/grid.css" rel="stylesheet">
<style>

.grid {
  grid-template-columns: 50px 100px;
  grid-template-rows: 50px 100px;
  width: 150px;
  height: 150px;
  /* Ensures that the grid container is the containing block of the absolutely positioned grid children. */
  position: relative;
  /* Ensures that the grid container is the containing block of the fixed positioned grid children. */
  transform: translate(10px, 20px);
}

.absolute {
  position: absolute;
}

.fixed {
  position: fixed;
}

.offsetLeft100 {
  left: 100px;
}

</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script type="text/javascript">
  setup({ explicit_done: true });
</script>

<body onload="document.fonts.ready.then(() => { checkLayout('.grid'); })">

<div id="log"></div>

<p>Absolutely positioned:</p>

<div class="grid">
  <div class="sizedToGridArea absolute autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea absolute autoRowAutoColumn offsetLeft100"
    data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea absolute autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea absolute autoRowAutoColumn offsetLeft100"
    data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
    <div class="sizedToGridArea absolute autoRowAutoColumn"
      data-offset-x="50" data-offset-y="0" data-expected-width="150" data-expected-height="150">
    </div>
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="50" data-expected-width="50" data-expected-height="100">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
    <div class="sizedToGridArea absolute autoRowAutoColumn offsetLeft100"
      data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
    </div>
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="50" data-expected-width="50" data-expected-height="100">
  </div>
</div>

<p>Fixed positioned:</p>

<div class="grid">
  <div class="sizedToGridArea fixed autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea fixed autoRowAutoColumn offsetLeft100"
    data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea fixed autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea fixed autoRowAutoColumn offsetLeft100"
    data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
    <div class="sizedToGridArea fixed autoRowAutoColumn"
      data-offset-x="50" data-offset-y="0" data-expected-width="150" data-expected-height="150">
    </div>
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="50" data-expected-width="50" data-expected-height="100">
  </div>
</div>

<div class="grid">
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="50" data-offset-y="0" data-expected-width="100" data-expected-height="50">
    <div class="sizedToGridArea fixed autoRowAutoColumn offsetLeft100"
      data-offset-x="100" data-offset-y="0" data-expected-width="150" data-expected-height="150">
    </div>
  </div>
  <div class="sizedToGridArea autoRowAutoColumn"
    data-offset-x="0" data-offset-y="50" data-expected-width="50" data-expected-height="100">
  </div>
</div>

</body>
